Add files via upload
Adjusted color in the app. It no longer has issues on the final page with the text whiting out and being difficult to read, as well as the bar at the top of the screen. The issue was caused by the app trying to adjust colors based on whether the device was in light or dark mode. The app is currently limited to 3 different types of rest. I am going to be looking into adding more to include all 7 types of rest.